+++ ultranest/stepfuncs.pyx
@@ -13,8 +13,8 @@ cimport cython
 from cython.parallel import prange
 
 
-ctypedef np.int32_t decl_int_t
-int_dtype = np.int32
+ctypedef np.int64_t decl_int_t
+int_dtype = np.int64
 
 
 @cython.boundscheck(False)
